Week 6 Day2 assignment 27

+++ Week5/Day5/Project_5/src/main/java/com/abdulrahman/project_5/Location/service/LocationService.java
@@ -36,6 +36,10 @@ public class LocationService {
         if (temp == null) {
             throw new ApiException("Not found");
         }
+        Location teacherAddress = teacherAddressRepo.findTeacherAddressById(id);
+        if (teacherAddress == null) {
+            throw new ApiException("No address found for the teacher");
+        }
         locationRepo.delete(temp);
 
     }
